Only use bus and no class for channel devices
It's wrong (and explicit oops with kernel >= 2.6.19) to set both
bus and class for the same device. But setting the class is not needed
in order to create a device. So just remove our dynamic devices from the
dahdi_class. It will only contain the fixed-named devices.

This fixes regression from cb4e4d006857d7fbc2ba51c63a96f2c97f2738d1.

Signed-off-by: Tzafrir Cohen <tzafrir.cohen@xorcom.com>